* refactoring

+++ src/main/java/com/neodem/rays/PaintableRay.java
@@ -11,21 +11,21 @@ import java.awt.Graphics;
  */
 public class PaintableRay implements Paintable {
 
-    private final int height;
+    private final float height;
     private final int locX;
     private final int locY;
     private final Color color;
 
-    public PaintableRay(int height, int locX, int midY, Color color) {
+    public PaintableRay(float height, int locX, int midY, Color color) {
         this.height = height;
         this.locX = locX;
-        this.locY = midY - (height/2);
+        this.locY = midY - ((int) height / 2);
         this.color = color;
     }
 
     @Override
     public void paint(Graphics g) {
         g.setColor(color.getAWTColor());
-        g.fillRect(locX, locY, 1, height);
+        g.fillRect(locX, locY, 1, (int) height);
     }
 }
